modified command prompt to display hostname:basename(crt_dir) (.bashrc file). build.sh: check if an image exists before doing anything else. cleanup.sh: remove an image only if found on disk. also simplify output. upload.sh: upload an image only if found on disk. simplify output.

+++ util/docker/open3d-test/tools/cleanup.sh
@@ -11,16 +11,23 @@ for ubuntu in ${ubuntu_version[@]}; do
     for bundle in ${bundle_type[@]}; do
         . ${REAL_PATH}/set_variables.sh ${ubuntu} ${bundle}
 
-        echo "removing ${IMAGE_NAME}..."
-        docker image rm ${IMAGE_NAME} >/dev/null 2>&1
+        # remove the image only if found locally
+        if [ 0 -eq ${IMAGE_EXISTS} ]; then
+            printf "removing ${IMAGE_NAME}..."
+            docker image rm ${IMAGE_NAME} >/dev/null 2>&1
+            printf "done\n"
+        fi
 
         for env in ${env_type[@]}; do
             . ${REAL_PATH}/set_variables.sh ${ubuntu} ${bundle} ${env}
 
-            echo "removing ${IMAGE_NAME}..."
-            docker image rm ${IMAGE_NAME} >/dev/null 2>&1
+            # remove the image only if found locally
+            if [ 0 -eq ${IMAGE_EXISTS} ]; then
+                printf "removing ${IMAGE_NAME}..."
+                docker image rm ${IMAGE_NAME} >/dev/null 2>&1
+                printf "done\n"
+            fi
         done
     done
 done
-
 echo
